refactor: break post-news action into module
