+++ native/cli/src/action.rs
@@ -1,0 +1,3 @@
+mod post_news;
+
+pub use post_news::PostNews;
